safari-600.7-branch/LayoutTests/fast/regions/cssom/get-region-flow-ranges-display-none-expected.txt

[CSSRegions] Implement Region::getRegionFlowRanges for elements with display:none (Range should not break ranges at display:none)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

1
3
5
Flow test
PASS range1.length is 1
PASS Array [wrapper,0,wrapper,1] is equal to [wrapper,0,wrapper,1]
PASS successfullyParsed is true

TEST COMPLETE
